Bluetooth: hci_core: temporarily disable sync train on device initialization
On radon's Bluetooth firmware, this somehow seems to lock it up entirely. It
might just be due to missing calibration data on our DVT units, or it might
be something else entirely. In any case, this gets it up and running.

This is just a temporary hack so we can ship, and we need to eventually get
to the bottom of it.